moved duplicated classes from test_sidechain*.cpp to tx_creation_utils.h
