+++ src/gtest/test_sidechain_blocks.cpp
@@ -12,58 +12,6 @@
 #include <miner.h>
 #include <gtest/libzendoo_test_files.h>
 
-class CInMemorySidechainDb final: public CCoinsView {
-public:
-    CInMemorySidechainDb()  = default;
-    virtual ~CInMemorySidechainDb() = default;
-
-    bool HaveSidechain(const uint256& scId) const override {
-        return sidechainsInMemoryMap.count(scId) && sidechainsInMemoryMap.at(scId).flag != CSidechainsCacheEntry::Flags::ERASED;
-    }
-    bool GetSidechain(const uint256& scId, CSidechain& info) const override {
-        if(!HaveSidechain(scId))
-            return false;
-        info = sidechainsInMemoryMap.at(scId).sidechain;
-        return true;
-    }
-
-    bool HaveSidechainEvents(int height)  const override {
-        return eventsInMemoryMap.count(height) && eventsInMemoryMap.at(height).flag != CSidechainEventsCacheEntry::Flags::ERASED;
-    }
-    bool GetSidechainEvents(int height, CSidechainEvents& scEvents) const override {
-        if(!HaveSidechainEvents(height))
-            return false;
-        scEvents = eventsInMemoryMap.at(height).scEvents;
-        return true;
-    }
-
-    virtual void GetScIds(std::set<uint256>& scIdsList) const override {
-        for (auto& entry : sidechainsInMemoryMap)
-            scIdsList.insert(entry.first);
-        return;
-    }
-
-    bool BatchWrite(CCoinsMap &mapCoins, const uint256 &hashBlock,
-                    const uint256 &hashAnchor, CAnchorsMap &mapAnchors,
-                    CNullifiersMap &mapNullifiers, CSidechainsMap& sidechainMap,
-                    CSidechainEventsMap& mapSidechainEvents, CCswNullifiersMap& cswNullifiers) override
-    {
-        for (auto& entryToWrite : sidechainMap)
-            WriteMutableEntry(entryToWrite.first, entryToWrite.second, sidechainsInMemoryMap);
-
-        for (auto& entryToWrite : mapSidechainEvents)
-            WriteMutableEntry(entryToWrite.first, entryToWrite.second, eventsInMemoryMap);
-
-        sidechainMap.clear();
-        mapSidechainEvents.clear();
-        return true;
-    }
-
-private:
-    mutable boost::unordered_map<uint256, CSidechainsCacheEntry, CCoinsKeyHasher> sidechainsInMemoryMap;
-    mutable boost::unordered_map<int, CSidechainEventsCacheEntry> eventsInMemoryMap;
-};
-
 class SidechainsConnectCertsBlockTestSuite : public ::testing::Test {
 public:
     SidechainsConnectCertsBlockTestSuite():
@@ -84,7 +32,7 @@ public:
         UnloadBlockIndex();
         mGlobalForkTips.clear();
 
-        fakeChainStateDb   = new CInMemorySidechainDb();
+        fakeChainStateDb   = new blockchain_test_utils::CInMemorySidechainDb();
         sidechainsView     = new txCreationUtils::CNakedCCoinsViewCache(fakeChainStateDb);
 
         dummyHash = dummyBlock.GetHash();
@@ -105,7 +53,7 @@ public:
     };
 
 protected:
-    CInMemorySidechainDb  *fakeChainStateDb;
+    blockchain_test_utils::CInMemorySidechainDb  *fakeChainStateDb;
     txCreationUtils::CNakedCCoinsViewCache *sidechainsView;
 
     //helpers
@@ -696,7 +644,7 @@ public:
 
         UnloadBlockIndex();
 
-        fakeChainStateDb   = new CInMemorySidechainDb();
+        fakeChainStateDb   = new blockchain_test_utils::CInMemorySidechainDb();
         blockchainView     = new CCoinsViewCache(fakeChainStateDb);
     };
 
@@ -711,7 +659,7 @@ public:
     };
 
 protected:
-    CInMemorySidechainDb *fakeChainStateDb;
+    blockchain_test_utils::CInMemorySidechainDb *fakeChainStateDb;
     CCoinsViewCache      *blockchainView;
 
     std::vector<TxPriority> vecPriority;
